1) исправлен баг сопределением ширины в TMCUFonts::getTextSizes 2) Компонент TComponentListVertical - контейнер для вертикального размещения элеметов. Выводит те что помещаются на экране. Выводит с учётом высоты предыдущего элемента. 3) в шрифтах ререшёл на u8/u16 стиль описания типов 4) витруальные методы view и getHeight

+++ MCU/Fonts/emb_fonts.h
@@ -1,17 +1,17 @@
 #ifndef EMB_FONTS_H
 #define EMB_FONTS_H
 
-#include <stdint.h>
+#include <stm32f4xx.h>
 
  typedef struct {
-     const int code;
-     const uint8_t* image;
-     const int width;
+     const u8 code;
+     const u8* image;
+     const u8 width;
      } tChar;
 
  typedef struct {
-     const int length;
-     const int height;
+     const u16 length;
+     const u8 height;
      const tChar *chars;
      } tFont;
